+++ tests/test_chatroom.py
@@ -16,6 +16,7 @@ import pytest
 
 from kaggle_benchmarks import actors, chats
 from kaggle_benchmarks.messages import Message
+from kaggle_benchmarks.rooms import ChatRoom, Participant
 from tests.mocks import MockedChat
 
 # ── Context Manager ──
@@ -25,31 +26,34 @@ def test_chatroom_rejects_duplicate_participants():
     """Same LLMChat added twice without name override is rejected (name collision)."""
     alice = MockedChat.from_contents(["x"], name="Alice")
 
-    room = chats.ChatRoom()
+    room = ChatRoom()
     room.add_participant(alice)
     with pytest.raises(ValueError, match="already taken"):
         room.add_participant(alice)
 
 
-def test_add_participant_same_llm_creates_distinct_clones():
-    """Same LLMChat object added with different names yields two distinct clones."""
+def test_add_participant_same_llm_creates_distinct_participants():
+    """Same LLMChat object added with different names yields two distinct Participants."""
     shared = MockedChat.from_contents(["x"], name="Model", cycle=True)
-    room = chats.ChatRoom()
+    room = ChatRoom()
     alice = room.add_participant(shared, name="Alice")
     bob = room.add_participant(shared, name="Bob")
 
     assert alice is not bob
-    assert alice is not shared
-    assert bob is not shared
+    assert isinstance(alice, Participant)
+    assert isinstance(bob, Participant)
+    # Both wrap the same backing actor
+    assert alice.actor is shared
+    assert bob.actor is shared
     assert alice.name == "Alice"
     assert bob.name == "Bob"
     assert len(room.participants) == 2
 
 
 def test_add_participant_same_llm_independent_perspectives():
-    """Clones from the same LLM have correctly separated perspectives."""
+    """Participants from the same LLM have correctly separated perspectives."""
     shared = MockedChat.from_contents(["hello", "world"], name="Model", cycle=True)
-    room = chats.ChatRoom()
+    room = ChatRoom()
     alice = room.add_participant(shared, name="Alice")
     bob = room.add_participant(shared, name="Bob")
 
@@ -58,7 +62,7 @@ def test_add_participant_same_llm_independent_perspectives():
         alice.reply()
         bob.reply()
 
-    # Messages attributed to correct clone
+    # Messages attributed to correct participant
     assert room.messages[1].sender is alice
     assert room.messages[2].sender is bob
 
@@ -77,40 +81,39 @@ def test_add_participant_rejects_duplicate_name():
     """Different LLMs with the same effective name are rejected."""
     a = MockedChat.from_contents(["x"], name="Alice")
     b = MockedChat.from_contents(["y"], name="Bob")
-    room = chats.ChatRoom()
+    room = ChatRoom()
     room.add_participant(a)
     with pytest.raises(ValueError, match="already taken"):
         room.add_participant(b, name="Alice")
 
 
-def test_add_participant_plain_actor_duplicate_rejected():
-    """Same plain Actor instance added twice is rejected."""
+def test_add_participant_plain_actor_rejected():
+    """Plain Actor instance added to room is rejected with TypeError."""
     game = actors.Actor(name="Game")
-    room = chats.ChatRoom()
-    room.add_participant(game)
-    with pytest.raises(ValueError):
-        room.add_participant(game, name="Host")
+    room = ChatRoom()
+    with pytest.raises(TypeError, match="requires an LLMChat instance"):
+        room.add_participant(game)
 
 
 def test_chatroom_sets_active_context():
     """with room: must make the ChatRoom the active chat."""
     alice = MockedChat.from_contents(["hi"], name="Alice", cycle=True)
-    room = chats.ChatRoom()
+    room = ChatRoom()
     alice = room.add_participant(alice)
     with room:
         assert chats.get_current_chat() is room
     assert chats.get_current_chat() is not room
 
 
-# ── say() / reply() Primitives ──
+# ── reply() Primitive ──
 
 
 def test_llmchat_reply_appends_to_room():
-    alice = MockedChat.from_contents(["I agree!"], name="Alice", cycle=True)
-    bob = MockedChat.from_contents(["x"], name="Bob")
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
+    alice_mock = MockedChat.from_contents(["I agree!"], name="Alice", cycle=True)
+    bob_mock = MockedChat.from_contents(["x"], name="Bob")
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
+    room.add_participant(bob_mock)
 
     with room:
         room.post("Topic: AI safety")
@@ -122,48 +125,20 @@ def test_llmchat_reply_appends_to_room():
     assert room.messages[1].content == "I agree!"
 
 
-def test_actor_say_appends_to_room():
-    game = actors.Actor(name="Game")
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    room = chats.ChatRoom()
-    game = room.add_participant(game)
-    alice = room.add_participant(alice)
-
-    with room:
-        game.say("Board: X|O|_")
-
-    assert len(room.messages) == 1
-    assert room.messages[0].sender is game
-    assert room.messages[0].content == "Board: X|O|_"
-
-
-def test_llmchat_reply_outside_room_raises():
-    alice = MockedChat.from_contents(["x"], name="Alice")
+def test_participant_reply_outside_room_raises():
+    """Participant.reply() must be called inside a ChatRoom context."""
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
     with pytest.raises(RuntimeError, match="ChatRoom context"):
         alice.reply()
 
 
-def test_actor_say_outside_room_raises():
-    game = actors.Actor(name="Game")
-    with pytest.raises(RuntimeError, match="ChatRoom context"):
-        game.say("hello")
-
-
-def test_actor_say_without_message_raises_typeerror():
-    """Actor.say() requires a message argument."""
-    game = actors.Actor(name="Game")
-    room = chats.ChatRoom()
-    game = room.add_participant(game)
-    with room:
-        with pytest.raises(TypeError):
-            game.say()
-
-
 def test_prompt_inside_chatroom_raises():
     """prompt() must not be called inside an active ChatRoom."""
-    alice = MockedChat.from_contents(["x"], name="Alice", cycle=True)
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice", cycle=True)
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
     with room:
         with pytest.raises(RuntimeError, match="cannot be called inside"):
             alice.prompt("hello")
@@ -173,9 +148,9 @@ def test_prompt_inside_chatroom_raises():
 
 
 def test_perspective_does_not_mutate_originals():
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
     original = Message(sender=alice, content="original")
     room.history = [original]
 
@@ -186,11 +161,11 @@ def test_perspective_does_not_mutate_originals():
 
 def test_perspective_filters_invisible_messages():
     """Messages with is_visible_to_llm=False are not shown in perspective."""
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    bob = MockedChat.from_contents(["x"], name="Bob")
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    bob_mock = MockedChat.from_contents(["x"], name="Bob")
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
+    bob = room.add_participant(bob_mock)
 
     room.history = [
         Message(sender=alice, content="I can be seen"),
@@ -205,13 +180,11 @@ def test_perspective_filters_invisible_messages():
 
 def test_perspective_preserves_avatars():
     """Synthetic actors in perspective should retain original avatars."""
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    alice.avatar = "🐺"
-    bob = MockedChat.from_contents(["x"], name="Bob")
-    bob.avatar = "🧙"
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    bob_mock = MockedChat.from_contents(["x"], name="Bob")
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock, avatar="🐺")
+    bob = room.add_participant(bob_mock, avatar="🧙")
 
     room.history = [
         Message(sender=alice, content="hello"),
@@ -229,22 +202,20 @@ def test_perspective_preserves_avatars():
 def test_system_prompt_composition():
     """System prompt contains roster, room instructions, personal prompt,
     narrator note, but NOT peers' system_prompts."""
-    alice = MockedChat.from_contents(["hello"], name="Alice")
-    alice.system_prompt = "SECRET: You are a werewolf"
-    bob = MockedChat.from_contents(["hi"], name="Bob")
-    bob.system_prompt = "SECRET: You are a villager"
+    alice_mock = MockedChat.from_contents(["hello"], name="Alice")
+    bob_mock = MockedChat.from_contents(["hi"], name="Bob")
 
-    room = chats.ChatRoom(
+    room = ChatRoom(
         system_prompt="A game of Werewolf.",
         name="Moderator",
     )
-    room.add_participant(alice)
-    room.add_participant(bob)
+    alice = room.add_participant(alice_mock, system_prompt="SECRET: You are a werewolf")
+    room.add_participant(bob_mock, system_prompt="SECRET: You are a villager")
 
     with room:
         alice.reply()
 
-    _, kwargs = alice.invocations[0]
+    _, kwargs = alice_mock.invocations[0]
     system = kwargs["system"]
 
     # Identity & roster
@@ -260,16 +231,11 @@ def test_system_prompt_composition():
     assert 'Note on "Moderator"' in system
 
 
-def test_system_prompt_via_from_contents():
-    alice = MockedChat.from_contents(["x"], name="Alice", system_prompt="argues FOR")
-    assert alice.system_prompt == "argues FOR"
-
-
 def test_post_sender_matches_room_name():
     """room.post() sender name matches the room's name."""
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    room = chats.ChatRoom(name="Moderator")
-    room.add_participant(alice)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    room = ChatRoom(name="Moderator")
+    room.add_participant(alice_mock)
     with room:
         msg = room.post("Hello")
     assert msg.sender.name == "Moderator"
@@ -279,11 +245,11 @@ def test_post_sender_matches_room_name():
 
 
 def test_visible_to_filters_messages():
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    bob = MockedChat.from_contents(["x"], name="Bob")
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    bob_mock = MockedChat.from_contents(["x"], name="Bob")
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
+    bob = room.add_participant(bob_mock)
 
     with room:
         room.post("Public message")
@@ -296,17 +262,17 @@ def test_visible_to_filters_messages():
 def test_private_channel_isolation_and_parent_visibility():
     """Covers: parent history visible from child, private content hidden
     from non-members, private messages tagged with channel name."""
-    alice = MockedChat.from_contents(["wolf plan", "I'm innocent!"], name="Alice")
-    bob = MockedChat.from_contents(["ack"], name="Bob")
-    charlie = MockedChat.from_contents(["suspicious"], name="Charlie")
+    alice_mock = MockedChat.from_contents(["wolf plan", "I'm innocent!"], name="Alice")
+    bob_mock = MockedChat.from_contents(["ack"], name="Bob")
+    charlie_mock = MockedChat.from_contents(["suspicious"], name="Charlie")
 
-    room = chats.ChatRoom(
+    room = ChatRoom(
         system_prompt="Werewolf game.",
         name="Moderator",
     )
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
-    charlie = room.add_participant(charlie)
+    alice = room.add_participant(alice_mock)
+    bob = room.add_participant(bob_mock)
+    charlie = room.add_participant(charlie_mock)
 
     with room:
         room.post("Day starts.")
@@ -322,14 +288,14 @@ def test_private_channel_isolation_and_parent_visibility():
         charlie.reply()  # "suspicious"
 
     # Charlie (non-member) must NOT see wolf chat content
-    charlie_msgs, _ = charlie.invocations[0]
+    charlie_msgs, _ = charlie_mock.invocations[0]
     charlie_content = " ".join(m.content for m in charlie_msgs)
     assert "wolf plan" not in charlie_content
     assert "Pick a victim" not in charlie_content
     assert "Day starts" in charlie_content
 
     # Bob (member) should see parent history + tagged private messages
-    bob_msgs, _ = bob.invocations[0]
+    bob_msgs, _ = bob_mock.invocations[0]
     bob_content = " ".join(m.content for m in bob_msgs)
     assert "Day starts" in bob_content  # parent visible from child
     private_tagged = [m for m in bob_msgs if "private: Wolf Night" in m.content]
@@ -338,24 +304,27 @@ def test_private_channel_isolation_and_parent_visibility():
 
 def test_private_channel_validates_participants():
     """private_channel() rejects participants not in the parent room."""
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    bob = MockedChat.from_contents(["x"], name="Bob")
-    outsider = MockedChat.from_contents(["x"], name="Outsider")
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    bob_mock = MockedChat.from_contents(["x"], name="Bob")
+    outsider_mock = MockedChat.from_contents(["x"], name="Outsider")
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
+    room.add_participant(bob_mock)
 
+    outsider_participant = Participant(
+        outsider_mock, name="Outsider", avatar="", role="assistant"
+    )
     with pytest.raises(ValueError, match="Unknown: Outsider"):
-        room.private_channel([alice, outsider], name="Bad Channel")
+        room.private_channel([alice, outsider_participant], name="Bad Channel")
 
 
 def test_private_channel_rejects_same_name_as_parent():
     """private_channel() rejects name matching the parent room's name."""
-    alice = MockedChat.from_contents(["x"], name="Alice")
-    bob = MockedChat.from_contents(["x"], name="Bob")
-    room = chats.ChatRoom(name="GameRoom")
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
+    alice_mock = MockedChat.from_contents(["x"], name="Alice")
+    bob_mock = MockedChat.from_contents(["x"], name="Bob")
+    room = ChatRoom(name="GameRoom")
+    alice = room.add_participant(alice_mock)
+    room.add_participant(bob_mock)
 
     with pytest.raises(ValueError, match="must differ from parent room name"):
         room.private_channel([alice], name="GameRoom")
@@ -363,14 +332,14 @@ def test_private_channel_rejects_same_name_as_parent():
 
 def test_nested_channels_three_levels():
     """Channel-in-channel: root -> team -> leader, with correct visibility."""
-    alice = MockedChat.from_contents(["intel", "report"], name="Alice")
-    bob = MockedChat.from_contents(["ack"], name="Bob")
-    charlie = MockedChat.from_contents(["cover"], name="Charlie")
+    alice_mock = MockedChat.from_contents(["intel", "report"], name="Alice")
+    bob_mock = MockedChat.from_contents(["ack"], name="Bob")
+    charlie_mock = MockedChat.from_contents(["cover"], name="Charlie")
 
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
-    charlie = room.add_participant(charlie)
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
+    bob = room.add_participant(bob_mock)
+    charlie = room.add_participant(charlie_mock)
 
     team = room.private_channel([alice, bob], name="Team")
 
@@ -388,7 +357,7 @@ def test_nested_channels_three_levels():
                 alice.reply()  # "report"
 
     # Alice (leader) saw all 3 levels
-    alice_msgs_r2, _ = alice.invocations[1]
+    alice_msgs_r2, _ = alice_mock.invocations[1]
     alice_content = " ".join(m.content for m in alice_msgs_r2)
     assert "Public" in alice_content
     assert "Team only" in alice_content
@@ -403,9 +372,9 @@ def test_nested_channels_three_levels():
 
 def test_reentrant_chatroom_loop():
     """Same private channel entered multiple times in a loop."""
-    alice = MockedChat.from_contents(["r1", "r2", "r3"], name="Alice", cycle=True)
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
+    alice_mock = MockedChat.from_contents(["r1", "r2", "r3"], name="Alice", cycle=True)
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
     channel = room.private_channel([alice], name="Night")
 
     with room:
@@ -424,9 +393,9 @@ def test_reentrant_chatroom_loop():
 
 
 def test_meta_chat_points_to_room():
-    alice = MockedChat.from_contents(["response"], name="Alice", cycle=True)
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
+    alice_mock = MockedChat.from_contents(["response"], name="Alice", cycle=True)
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
     with room:
         alice.reply()
     assert room.messages[0]._meta.get("chat") is room
@@ -467,9 +436,11 @@ class MockedLLMResponseChat(actors.LLMChat):
 def test_meta_llm_response_path():
     """Verify _meta contains LLMResponse.meta fields (usage tokens etc.)
     when the LLMResponse code path is used."""
-    alice = MockedLLMResponseChat.from_contents(["hello"], name="Alice", cycle=True)
-    room = chats.ChatRoom()
-    alice = room.add_participant(alice)
+    alice_mock = MockedLLMResponseChat.from_contents(
+        ["hello"], name="Alice", cycle=True
+    )
+    room = ChatRoom()
+    alice = room.add_participant(alice_mock)
 
     with room:
         alice.reply()
@@ -486,20 +457,19 @@ def test_meta_llm_response_path():
 def test_three_player_perspective():
     """Verify exact messages, roles, prefixes, and system prompts that
     each of 3 LLMs receives across 2 rounds (6 invocations total)."""
-    alice = MockedChat.from_contents(["Alice R1", "Alice R2"], name="Alice")
-    alice.system_prompt = "I am Alice"
-    bob = MockedChat.from_contents(["Bob R1", "Bob R2"], name="Bob")
-    bob.system_prompt = "I am Bob"
-    charlie = MockedChat.from_contents(["Charlie R1", "Charlie R2"], name="Charlie")
-    charlie.system_prompt = "I am Charlie"
+    alice_mock = MockedChat.from_contents(["Alice R1", "Alice R2"], name="Alice")
+    bob_mock = MockedChat.from_contents(["Bob R1", "Bob R2"], name="Bob")
+    charlie_mock = MockedChat.from_contents(
+        ["Charlie R1", "Charlie R2"], name="Charlie"
+    )
 
-    room = chats.ChatRoom(
+    room = ChatRoom(
         system_prompt="A three-way discussion.",
         name="Moderator",
     )
-    alice = room.add_participant(alice)
-    bob = room.add_participant(bob)
-    charlie = room.add_participant(charlie)
+    alice = room.add_participant(alice_mock, system_prompt="I am Alice")
+    bob = room.add_participant(bob_mock, system_prompt="I am Bob")
+    charlie = room.add_participant(charlie_mock, system_prompt="I am Charlie")
 
     with room:
         room.post("Topic: Testing")
@@ -513,7 +483,7 @@ def test_three_player_perspective():
     assert len(room.messages) == 7
 
     # ── Alice R1: sees [post] ──
-    a1_msgs, a1_kw = alice.invocations[0]
+    a1_msgs, a1_kw = alice_mock.invocations[0]
     assert len(a1_msgs) == 1
     assert a1_msgs[0].sender.role == "user"
     assert "[Moderator]: Topic: Testing" in a1_msgs[0].content
@@ -522,7 +492,7 @@ def test_three_player_perspective():
     assert "I am Bob" not in a1_kw["system"]
 
     # ── Alice R2: sees [post, self(assistant), Bob, Charlie] ──
-    a2_msgs, _ = alice.invocations[1]
+    a2_msgs, _ = alice_mock.invocations[1]
     assert len(a2_msgs) == 4
     assert a2_msgs[1].sender.role == "assistant"
     assert a2_msgs[1].content == "Alice R1"
@@ -531,7 +501,7 @@ def test_three_player_perspective():
     assert "[Charlie]: Charlie R1" in a2_msgs[3].content
 
     # ── Bob R1: sees [post, Alice R1] ──
-    b1_msgs, b1_kw = bob.invocations[0]
+    b1_msgs, b1_kw = bob_mock.invocations[0]
     assert len(b1_msgs) == 2
     assert b1_msgs[1].sender.role == "user"
     assert "[Alice]: Alice R1" in b1_msgs[1].content
@@ -539,14 +509,14 @@ def test_three_player_perspective():
     assert "I am Alice" not in b1_kw["system"]
 
     # ── Bob R2: sees [post, Alice R1, self(assistant), Charlie R1, Alice R2] ──
-    b2_msgs, _ = bob.invocations[1]
+    b2_msgs, _ = bob_mock.invocations[1]
     assert len(b2_msgs) == 5
     assert b2_msgs[2].sender.role == "assistant"
     assert b2_msgs[2].content == "Bob R1"
     assert "[Bob]:" not in b2_msgs[2].content
 
     # ── Charlie R2: sees all 6 prior messages ──
-    c2_msgs, _ = charlie.invocations[1]
+    c2_msgs, _ = charlie_mock.invocations[1]
     assert len(c2_msgs) == 6
     assert c2_msgs[3].sender.role == "assistant"
     assert c2_msgs[3].content == "Charlie R1"
